+++ v8go.cc
@@ -202,24 +202,294 @@ void ContextDispose(ContextPtr ptr) {
 
 /********** Value **********/
 
+#define LOCAL_VALUE(ptr)                           \
+  m_value* val = static_cast<m_value*>(ptr);       \
+  m_ctx* ctx = val->ctx_ptr;                       \
+  Isolate* iso = ctx->iso;                         \
+  Locker locker(iso);                              \
+  Isolate::Scope isolate_scope(iso);               \
+  HandleScope handle_scope(iso);                   \
+  Context::Scope context_scope(ctx->ptr.Get(iso)); \
+  Local<Value> value = val->ptr.Get(iso);
+
 void ValueDispose(ValuePtr ptr) {
   delete static_cast<m_value*>(ptr);
 }
 
 const char* ValueToString(ValuePtr ptr) {
-  m_value* val = static_cast<m_value*>(ptr);
-  m_ctx* ctx = val->ctx_ptr;
-  Isolate* iso = ctx->iso;
-
-  Locker locker(iso);
-  Isolate::Scope isolate_scope(iso);
-  HandleScope handle_scope(iso);
-  Context::Scope context_scope(ctx->ptr.Get(iso));
-
-  Local<Value> value = val->ptr.Get(iso);
+  LOCAL_VALUE(ptr);
   String::Utf8Value utf8(iso, value);
-
   return CopyString(utf8);
+}
+
+int ValueIsUndefined(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsUndefined();
+}
+
+int ValueIsNull(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsNull();
+}
+
+int ValueIsNullOrUndefined(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsNullOrUndefined();
+}
+
+int ValueIsTrue(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsTrue();
+}
+
+int ValueIsFalse(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsFalse();
+}
+
+int ValueIsName(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsName();
+}
+
+int ValueIsString(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsString();
+}
+
+int ValueIsSymbol(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsSymbol();
+}
+
+int ValueIsFunction(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsFunction();
+}
+
+int ValueIsObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsObject();
+}
+
+int ValueIsBigInt(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsBigInt();
+}
+
+int ValueIsBoolean(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsBoolean();
+}
+
+int ValueIsNumber(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsNumber();
+}
+
+int ValueIsExternal(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsExternal();
+}
+
+int ValueIsInt32(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsInt32();
+}
+
+int ValueIsUint32(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsUint32();
+}
+
+int ValueIsDate(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsDate();
+}
+
+int ValueIsArgumentsObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsArgumentsObject();
+}
+
+int ValueIsBigIntObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsBigIntObject();
+}
+
+int ValueIsNumberObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsNumberObject();
+}
+
+int ValueIsStringObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsStringObject();
+}
+
+int ValueIsSymbolObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsSymbolObject();
+}
+
+int ValueIsNativeError(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsNativeError();
+}
+
+int ValueIsRegExp(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsRegExp();
+}
+
+int ValueIsAsyncFunction(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsAsyncFunction();
+}
+
+int ValueIsGeneratorFunction(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsGeneratorFunction();
+}
+
+int ValueIsGeneratorObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsGeneratorObject();
+}
+
+int ValueIsPromise(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsPromise();
+}
+
+int ValueIsMap(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsMap();
+}
+
+int ValueIsSet(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsSet();
+}
+
+int ValueIsMapIterator(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsMapIterator();
+}
+
+int ValueIsSetIterator(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsSetIterator();
+}
+
+int ValueIsWeakMap(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsWeakMap();
+}
+
+int ValueIsWeakSet(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsWeakSet();
+}
+
+int ValueIsArray(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsArray();
+}
+
+int ValueIsArrayBuffer(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsArrayBuffer();
+}
+
+int ValueIsArrayBufferView(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsArrayBufferView();
+}
+
+int ValueIsTypedArray(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsTypedArray();
+}
+
+int ValueIsUint8Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsUint8Array();
+}
+
+int ValueIsUint8ClampedArray(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsUint8ClampedArray();
+}
+
+int ValueIsInt8Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsInt8Array();
+}
+
+int ValueIsUint16Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsUint16Array();
+}
+
+int ValueIsInt16Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsInt16Array();
+}
+
+int ValueIsUint32Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsUint32Array();
+}
+
+int ValueIsInt32Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsInt32Array();
+}
+
+int ValueIsFloat32Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsFloat32Array();
+}
+
+int ValueIsFloat64Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsFloat64Array();
+}
+
+int ValueIsBigInt64Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsBigInt64Array();
+}
+
+int ValueIsBigUint64Array(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsBigUint64Array();
+}
+
+int ValueIsDataView(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsDataView();
+}
+
+int ValueIsSharedArrayBuffer(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsSharedArrayBuffer();
+}
+
+int ValueIsProxy(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsProxy();
+}
+
+int ValueIsWasmModuleObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsWasmModuleObject();
+}
+
+int ValueIsModuleNamespaceObject(ValuePtr ptr) {
+  LOCAL_VALUE(ptr);
+  return value->IsModuleNamespaceObject();
 }
 
 /********** Version **********/