+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Value methods for checking value kind (is string, number, array etc)
 - C formatting via `clang-format` to aid future development
 
 ## [v0.3.0]